+++ h5netcdf/attrs.py
@@ -22,8 +22,15 @@ class Attributes(MutableMapping):
         self._check_dtype = check_dtype
 
     def __getitem__(self, key):
+        import h5py
+
         if key in _HIDDEN_ATTRS:
             raise KeyError(key)
+        # see https://github.com/h5netcdf/h5netcdf/issues/94 for details
+        if isinstance(self._h5attrs[key], h5py.Empty):
+            string_info = h5py.check_string_dtype(self._h5attrs[key].dtype)
+            if string_info and string_info.length == 1:
+                return b""
         return self._h5attrs[key]
 
     def __setitem__(self, key, value):